requirements.txt
crewai 
streamlit 
pandas 
langchain_community 
langchain_groq
matplotlib 
numpy 
pillow 
crewai[tools]==0.30.11
python-dotenv